fix(personas): целочисленное деление в topSql — business/smm/personal не классифицировались
В запросе топ-ассистентов по персоне доля категорий считалась целочисленно
(m.biz / m.total, оба bigint) → всегда 0 → business/smm/personal_growth не
попадали в свою персону и проваливались в mixed. Из-за этого у Business пустой
breakdown ассистентов, а Mixed «съедал» их. Каст ::numeric (как в главном
запросе персон). Чинит и вкладку «Персоны», и assistant_breakdown в снапшоте.

+++ src/monitoring/product/personas.service.ts
@@ -219,9 +219,9 @@ export class PersonasService implements OnModuleInit {
           m.user_id,
           CASE
             WHEN COALESCE(g.gen_count, 0) >= 5                     THEN 'content_creator'
-            WHEN m.smm / NULLIF(m.total, 0) >= 0.40                THEN 'smm'
-            WHEN m.biz / NULLIF(m.total, 0) >= 0.40                THEN 'business'
-            WHEN m.per / NULLIF(m.total, 0) >= 0.40                THEN 'personal_growth'
+            WHEN m.smm::numeric / NULLIF(m.total, 0) >= 0.40       THEN 'smm'
+            WHEN m.biz::numeric / NULLIF(m.total, 0) >= 0.40       THEN 'business'
+            WHEN m.per::numeric / NULLIF(m.total, 0) >= 0.40       THEN 'personal_growth'
             WHEN m.total < 5                                       THEN 'curious'
             ELSE 'mixed'
           END AS persona
